requirements.txt
google-generativeai>=0.3.0
streamlit>=1.24.0
streamlit-feedback>=0.1.0
streamlit-extras>=0.3.0
streamlit-card>=0.0.2
pandas>=2.0.0
pydeck>=0.8.0
python-dotenv>=1.0.0
geopy>=2.3.0
folium>=0.14.0
streamlit-folium>=0.15.1